% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-co2_df_short.R
\docType{data}
\name{co2_df_short}
\alias{co2_df_short}
\title{CO2 concentration}
\format{
A tibble with 1801 rows and 5 variables
\describe{
\item{datetime}{Datetime at which CO2 concentration was recorded.}
\item{temp_air}{Air temperature inside the flux chamber in Celsius.}
\item{temp_soil}{Ground temperature inside the flux chamber in Celsius.}
\item{conc}{CO2 concentration in ppm.}
\item{PAR}{Photosyntethically active radiation inside the chamber in micromol/s/sqm.}
}
}
\usage{
co2_df_short
}
\description{
Continous CO2 concentration as measured on the field
}
\examples{
co2_df_short
}
\keyword{datasets}
