% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-co2_fluxes_lrc.R
\docType{data}
\name{co2_fluxes_lrc}
\alias{co2_fluxes_lrc}
\title{CO2 fluxes with PAR values}
\format{
A tibble with 257 rows and 5 variables
\describe{
\item{f_flux}{CO2 flux in mmol/sqm/hour.}
\item{datetime}{Time and date of the measurement.}
\item{PAR_ave}{Photosynthetically active radiation inside the chamber in
micromol/s/sqm averaged over the flux measurement.}
\item{type}{Type of measurement: ecosystems respiration (ER), net ecosystem
exchange (NEE), or light response curve (LRC).}
\item{warming}{Treatment: control or warming.}
}
}
\usage{
co2_fluxes_lrc
}
\description{
CO2 fluxes with photosynthetically active radiation (PAR) for testing and
examples of light response curves (LRC)
}
\examples{
co2_fluxes_lrc
}
\keyword{datasets}
