% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-co2_liahovden.R
\docType{data}
\name{co2_liahovden}
\alias{co2_liahovden}
\title{CO2 concentration at Liahovden}
\format{
A tibble with 89692 rows and 5 variables
\describe{
\item{datetime}{Datetime at which CO2 concentration was recorded.}
\item{temp_air}{Air temperature inside the flux chamber in Celsius.}
\item{temp_soil}{Ground temperature inside the flux chamber in Celsius.}
\item{conc}{CO2 concentration in ppm.}
\item{PAR}{Photosynthetically active radiation inside the chamber
in micromol/s/sqm.}
}
}
\usage{
co2_liahovden
}
\description{
CO2 concentration at Liahovden site, used in example in readme file
}
\examples{
co2_liahovden
}
\keyword{datasets}
