% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_calc_frac.R
\name{flux_calc_frac}
\alias{flux_calc_frac}
\title{for flux calculation in fractional units}
\usage{
flux_calc_frac(
  slope_df,
  slope_col,
  setup_volume,
  plot_area,
  f_fluxid,
  atm_pressure
)
}
\arguments{
\item{slope_col}{column containing the slope to calculate the flux}

\item{setup_volume}{volume of the flux chamber and instrument together in L,
can also be a column in case it is a variable}

\item{plot_area}{area of the plot in m^2,
can also be a column in case it is a variable}

\item{f_fluxid}{column containing the flux IDs}

\item{atm_pressure}{atmospheric pressure in atm,
can be a constant (numerical) or a variable (column name)}

\item{slopes_df}{dataframe of flux slopes}
}
\value{
df with fluxes
}
\description{
calculates fluxes with avg temp and pressure
}
\keyword{internal}
