% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_calc_vol.R
\name{flux_calc_vol}
\alias{flux_calc_vol}
\title{for flux calculation in volumetric units}
\usage{
flux_calc_vol(slopes_df, slope_col, setup_volume, plot_area, f_fluxid)
}
\arguments{
\item{slopes_df}{dataframe of flux slopes}

\item{slope_col}{column containing the slope to calculate the flux}

\item{setup_volume}{volume of the flux chamber and instrument together in L,
can also be a column in case it is a variable}

\item{plot_area}{area of the plot in m^2,
can also be a column in case it is a variable}

\item{f_fluxid}{column containing the flux IDs}
}
\value{
df with fluxes
}
\description{
calculates fluxes with instant temp and pressure
}
\keyword{internal}
