% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_check_item.R
\name{flux_check_item}
\alias{flux_check_item}
\title{check the items inside \code{flux_fun_check}}
\usage{
flux_check_item(arg, fn, msg, narg, df_name = NA)
}
\arguments{
\item{arg}{argument to be checked by \code{fn}}

\item{fn}{function to check \code{arg}}

\item{msg}{message to display in case \code{arg} is the wrong class}

\item{narg}{name of \code{arg}}

\item{df_name}{name of \code{arg} in case it is a data frame}
}
\description{
check the items inside \code{flux_fun_check}
}
\author{
Adam Klimes
}
