% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_cut.R
\name{flux_cut}
\alias{flux_cut}
\title{filter cut data before calculating fluxes}
\usage{
flux_cut(slope_df, cut_col, keep_filter)
}
\arguments{
\item{slope_df}{dataset containing slopes and cut column}

\item{cut_col}{column containing cutting information}

\item{keep_filter}{name in cut_col of data to keep}
}
\description{
filter cut data before calculating fluxes
}
