% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fitting_segment.R
\name{flux_fitting_segment}
\alias{flux_fitting_segment}
\title{Fitting function while accounting for leaks}
\usage{
flux_fitting_segment(
  conc_df,
  start_cut,
  end_cut,
  par_col,
  h2o_col,
  sign_str_col,
  h2o_correction,
  min_seg_length
)
}
\arguments{
\item{conc_df}{dataframe of gas concentration over time}

\item{start_cut}{time to discard at the start of the measurements
(in seconds)}

\item{end_cut}{time to discard at the end of the measurements (in seconds)}

\item{par_col}{column containing PAR data (segments; optional).}

\item{h2o_col}{column containing water vapour concentration
(segments; optional).}

\item{sign_str_col}{column containing signal strength
(segments; optional).}

\item{h2o_correction}{Should the gas concentration be corrected
for water vapor? (logical; segments)}

\item{min_seg_length}{minimum length (in seconds) for each segments
(segments). Note that the minimum segment length cannot be longer than half
the length of the measurements.}
}
\value{
a df with modeled gas concentration in linear segments, with slope,
r squared, adjusted r squared, p-value, average PAR (if provided) and
signal strength (if provided) for each segments.
}
\description{
separate measurements in linear segments segments
}
