% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_flag_count.R
\name{flux_flag_count}
\alias{flux_flag_count}
\title{Counts quality flags}
\usage{
flux_flag_count(
  flags_df,
  f_fluxid = f_fluxid,
  f_quality_flag = f_quality_flag,
  f_flags = c("ok", "discard", "zero", "force_discard", "start_error", "no_data",
    "force_ok", "force_zero", "force_lm", "no_slope"),
  show_total = TRUE
)
}
\arguments{
\item{flags_df}{dataframe of flux slopes}

\item{f_fluxid}{column containing fluxes unique ID}

\item{f_quality_flag}{column containing the quality flags}

\item{f_flags}{list of flags used in the dataset
(if different from default from flux_quality).
If not provided, it will list only the flags that are
present in the dataset (no showing 0).}

\item{show_total}{logical, if TRUE (default), adds a row with the total
number of fluxes}
}
\value{
a dataframe with the number of fluxes for each quality flags
and their proportion to the total
}
\description{
Provides a table of how many fluxes were attributed
which quality flag. This function is incorporated in
\link[fluxible]{flux_quality} as a message, but can be used alone to extract
a dataframe with the flag count.
}
\examples{
data(co2_conc)
slopes <- flux_fitting(co2_conc, conc, datetime, fit_type = "exp_zhao18")
slopes_flag <- flux_quality(slopes, conc)
flux_flag_count(slopes_flag)
}
\author{
Vincent Belde
}
