% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fortify.R
\name{flux_fortify}
\alias{flux_fortify}
\title{Fortify fluxes for visual evaluation}
\usage{
flux_fortify(
  slopes_df,
  f_conc = f_conc,
  f_datetime = f_datetime,
  f_ylim_upper = 800,
  f_ylim_lower = 400,
  f_facetid = "f_fluxid",
  y_text_position = 500
)
}
\arguments{
\item{slopes_df}{dataset containing slopes,
with flags produced by \link[fluxible:flux_quality]{flux_quality}}

\item{f_conc}{column with gas concentration}

\item{f_datetime}{column with datetime of each data point}

\item{f_ylim_upper}{y axis upper limit}

\item{f_ylim_lower}{y axis lower limit}

\item{f_facetid}{character vector of columns to use as facet IDs. Note that
they will be united, and that has to result in a unique facet ID for each
measurement. Default is \code{f_fluxid}}
}
\description{
Prepares the fluxes for plotting.
}
\keyword{internal}
