% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fortify_exp.R
\name{flux_fortify_exp}
\alias{flux_fortify_exp}
\title{Fortify fluxes with exponential fit}
\usage{
flux_fortify_exp(slopes_df, f_datetime, kappamax)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{f_datetime}{column with datetime of each data point}

\item{kappamax}{indicating if kappamax was applied}
}
\description{
plots the fluxes that were fitted with
an exponential model
}
\keyword{internal}
