% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fortify_lin.R
\name{flux_fortify_lin}
\alias{flux_fortify_lin}
\title{Fortify fluxes with linear fit}
\usage{
flux_fortify_lin(slopes_df, f_datetime)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{f_datetime}{column with datetime of each data point}
}
\description{
Extracts fluxes that were fitted
with a linear model
}
\keyword{internal}
