% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fortify_quadratic.R
\name{flux_fortify_quadratic}
\alias{flux_fortify_quadratic}
\title{Fortify fluxes with a quadratic fit}
\usage{
flux_fortify_quadratic(slopes_df, f_datetime)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{f_datetime}{column with datetime of each data point}
}
\description{
specific part of flux_plot for
quadratic fit
}
\keyword{internal}
