% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_graph_lin.R
\name{flux_graph_lin}
\alias{flux_graph_lin}
\title{Graphing fluxes for fit evaluation}
\usage{
flux_graph_lin(
  slopes_df,
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  cut_col = "f_conc",
  fit_col = "f_fit",
  quality_flag_col = "f_quality_flag",
  fluxID_col = "f_fluxID",
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  start_col = "f_start",
  f_date_breaks = "1 min",
  f_minor_breaks = "10 sec",
  f_date_labels = "\%e/\%m \\n \%H:\%M",
  f_ylim_upper = 800,
  f_ylim_lower = 400,
  f_scales = "free",
  f_plotname = "plot_quality_lin.pdf",
  f_ncol = 5,
  f_nrow = 4,
  print_plot = "FALSE"
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}
}
\description{
graphs the fluxes and indicates what should be discarded or replaced by zero
}
\examples{
data(slopes0lin_flag)
flux_graph_lin(slopes0lin_flag, datetime_col = "datetime", conc_col = "conc", cut_col = "cut", fit_col = "fit", start_col = "start", print_plot = TRUE)
}
