% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_match_col.R
\name{flux_match_col}
\alias{flux_match_col}
\title{Using an already existing end column to slice measurements}
\usage{
flux_match_col(field_record, start_col, end_col, name_field_record)
}
\arguments{
\item{field_record}{dataframe recording which measurement happened when.
Has to contain at least a column containing the start of each measurement,
and any other column identifying the measurements.}

\item{start_col}{start column in field_record (\code{ymd_hms} format)}

\item{end_col}{end column in field_record (\code{ymd_hms} format)}

\item{name_field_record}{name of the df (for error message)}
}
\description{
Provides the \code{f_end} column for \code{flux_match}
}
\keyword{internal}
