% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_match_fixed.R
\name{flux_match_fixed}
\alias{flux_match_fixed}
\title{Using a fixed measurement length to slice the measurements}
\usage{
flux_match_fixed(field_record, start_col, measurement_length)
}
\arguments{
\item{field_record}{dataframe recording which measurement happened when.
Has to contain at least a column containing the start of each measurement,
and any other column identifying the measurements.}

\item{start_col}{start column in field_record (\code{ymd_hms} format)}

\item{measurement_length}{length of the measurement (in seconds)
from the start specified in the \code{field_record}}
}
\description{
Provides the \code{f_end} column for \code{flux_match}
}
