% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_param_lm.R
\name{flux_param_lm}
\alias{flux_param_lm}
\title{prepares text to print in flux_plot}
\usage{
flux_param_lm(slopes_df, cut_arg = "cut")
}
\arguments{
\item{slopes_df}{the slopes_df that is being provided to flux_plot}

\item{cut_arg}{argument pointing rows to be cut from the measurements}
}
\description{
creates a df with quality flags and quality diagnostics
to print on the plots produced by flux_plot.
flux_param_lm is for fit in the lm family (linear and quadratic)
flux_param_exp is for the exponential fit
}
