% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_param_qua.R
\name{flux_param_qua}
\alias{flux_param_qua}
\title{prepares text to print in flux_plot}
\usage{
flux_param_qua(slopes_df, conc_col)
}
\arguments{
\item{slopes_df}{the slopes_df that is being provided to flux_plot}

\item{conc_col}{column with gas concentration}
}
\description{
creates a df with quality flags and quality diagnostics
to print on the plots produced by flux_plot.
flux_param_lm is for fit in the lm family (linear and quadratic)
flux_param_exp is for the exponential fit
}
