% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_lin.R
\name{flux_plot_lin}
\alias{flux_plot_lin}
\title{plotting fluxes with linear fit}
\usage{
flux_plot_lin(slopes_df, f_conc, f_datetime, x_nudge, y_nudge)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{f_conc}{column with gas concentration}

\item{f_datetime}{column with datetime of each data point}

\item{x_nudge}{x nudge for position_nudge in geom_text}

\item{y_nudge}{y nudge for position_nudge in geom_text}
}
\description{
plots the fluxes that were fitted
with a linear model
}
