% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_longpdf.R
\name{flux_plot_longpdf}
\alias{flux_plot_longpdf}
\title{print plots in a single long pdf}
\usage{
flux_plot_longpdf(f_plot, f_plotname, nb_fluxid, longpdf_args)
}
\arguments{
\item{f_plot}{plot from flux_plots}

\item{f_plotname}{name of plot}

\item{nb_fluxid}{number of facets}

\item{longpdf_args}{arguments for plotly in the form
\code{list(ncol, width (cm), ratio)}}
}
\description{
print plots in a single long pdf
}
