% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_pdf.R
\name{flux_plot_pdf}
\alias{flux_plot_pdf}
\title{print plots in pdf}
\usage{
flux_plot_pdf(f_plot, f_plotname, plot_pages, facet_wrap_args)
}
\arguments{
\item{f_plot}{plot from flux_plot}

\item{f_plotname}{name of plot}

\item{plot_pages}{number of pages}

\item{facet_wrap_args}{list of arguments for
\link[ggforce:facet_wrap_paginate]{facet_wrap_paginate}}
}
\description{
print plots in pdf
}
