% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_plotly.R
\name{flux_plot_plotly}
\alias{flux_plot_plotly}
\title{print plots in pdf}
\usage{
flux_plot_plotly(f_plot, f_plotname, nb_fluxid, plotly_args)
}
\arguments{
\item{f_plot}{plot from flux_plot}

\item{f_plotname}{name of plot}

\item{facet_wrap_args}{list of arguments for
\link[ggforce:facet_wrap_paginate]{facet_wrap_paginate}}
}
\description{
print plots in pdf
}
