% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_segment.R
\name{flux_plot_segment}
\alias{flux_plot_segment}
\title{plots fluxes fitted with segmentation tool}
\usage{
flux_plot_segment(slopes_df, y_text_position, cut_arg)
}
\description{
provides plots for measurements that were fitted using
the segment model in flux_fitting
}
