% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_print_plot.R
\name{flux_print_plot}
\alias{flux_print_plot}
\title{Printing plot}
\usage{
flux_print_plot(f_plot, facet_wrap_args)
}
\arguments{
\item{f_plot}{plot from flux_plot}

\item{facet_wrap_args}{list of arguments for
\link[ggplot2:facet_wrap]{facet_wrap}}
}
\description{
printing plot from flux_plot
}
\keyword{internal}
