% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_exp.R
\name{flux_quality_exp}
\alias{flux_quality_exp}
\title{assessing fluxes quality calculated with an exponential fit}
\usage{
flux_quality_exp(
  slopes_df,
  fluxID_col = "f_fluxID",
  conc_col = "f_conc",
  b_col = "f_b",
  time_col = "f_time",
  fit_col = "f_fit",
  slope_col = "f_slope_tz",
  weird_fluxesID = c(),
  RMSE_threshold = 25,
  cor_threshold = 0.5,
  b_threshold = 1,
  ambient_conc = 421,
  error = 100
)
}
\description{
assessing fluxes quality calculated with an exponential fit
}
