% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_lm.R
\name{flux_quality_lm}
\alias{flux_quality_lm}
\title{quality assessment for the slopes estimated by flux_fitting}
\usage{
flux_quality_lm(
  slopes_df,
  conc_col,
  fluxid_col,
  slope_col,
  time_col,
  cut_col,
  pvalue_col,
  rsquared_col,
  force_discard = c(),
  force_ok = c(),
  ratio_threshold = 0,
  fit_type = c(),
  ambient_conc = 421,
  error = 100,
  pvalue_threshold = 0.3,
  rsquared_threshold = 0.7,
  cut_arg = "cut",
  name_df
)
}
\arguments{
\item{slopes_df}{dataset containing slopes, fluxID, p.value and r.squared}

\item{conc_col}{column with gas concentration}

\item{pvalue_col}{column containing the p-value of each flux}

\item{rsquared_col}{column containing the r squared to be used for
the quality assessment}

\item{force_discard}{vector of fluxIDs that should be discarded
by the user's decision}

\item{force_ok}{vector of fluxIDs for which the user wants to keep
the calculated slope despite a bad quality flag}

\item{pvalue_threshold}{threshold of p-value below which the change
of gas concentration over time is considered not significant (user decided)}

\item{rsquared_threshold}{threshold of r squared value below which
the linear model is considered an unsatisfactory fit}
}
\value{
same dataframe with added flag and corrected slopes columns
}
\description{
indicates if fluxes should be discarded or replaced by 0
according to parameters set by user.
flux_quality_lm is for the model of the lm family.
flux_quality_exp is for the exponential model.
}
