% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_segment.R
\name{flux_quality_segment}
\alias{flux_quality_segment}
\title{Quality control for segmented measurements}
\usage{
flux_quality_segment(
  slopes_df,
  pvalue_col,
  rsquared_col,
  f_flag_fit_col,
  par_col,
  sign_str_col,
  par_threshold,
  sign_str_threshold,
  pvalue_threshold,
  rsquared_threshold,
  sd_threshold,
  cut_arg,
  force_discard,
  force_ok
)
}
\description{
provides quality flags for the segments and entire measurements following the same standards as for the linear model
}
