% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_flux}
\alias{match_flux}
\title{Matching continously measured fluxes with measurement IDs}
\usage{
match_flux(raw_flux, field_record, startcrop = 10, measurement_length = 210)
}
\description{
Matching continously measured fluxes with measurement IDs
}
