% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pftc7_long.R
\docType{data}
\name{pftc7_long}
\alias{pftc7_long}
\title{CO2 concentration}
\format{
\describe{
\item{}{}
}
}
\usage{
pftc7_long
}
\description{
CO2 concentration with measurements meta data, measured with a tent setup
}
\examples{
pftc7_long
}
\keyword{datasets}
