% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pftc7_short.R
\docType{data}
\name{pftc7_short}
\alias{pftc7_short}
\title{CO2 concentration}
\format{
a tibble with 11 columns and 1665 rows
\describe{
\item{co2_conc}{CO2 concentration in ppm.}
\item{h2o_conc}{Water vapour concentration.}
\item{temperature_c}{Air temperature inside the flux chamber in Celsius.}
\item{pressure}{Pressure inside the chamber in atm.}
\item{signal_strength}{Strength of the signal in the sensor.}
\item{date_time}{Datetime at which CO2 concentration was recorded.}
\item{start_time}{Datetime at which the measurement was started.}
\item{file_name}{Name of individual flux file.}
\item{plot_id}{Unique ID for each plot.}
\item{par}{Photosynthetically active radiation inside the chamber
in micromol/s/sqm.}
\item{f_end}{Datetime at which the measurement ended.}
}
}
\usage{
pftc7_short
}
\description{
CO2 concentration with measurements meta data, measured with a tent setup
}
\examples{
pftc7_short
}
\keyword{datasets}
