% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-raw_terra.R
\docType{data}
\name{raw_terra}
\alias{raw_terra}
\title{CO2 and CH4 concentration}
\format{
A tibble with 21681 rows and 4 variables
\describe{
\item{co2_conc}{CO2 concentration in ppm}
\item{ch4_conc}{CH4 concentration in ppb}
\item{datetime}{Datetime on the datapoint}
\item{temp_air}{Air temperature inside the chamber in Celsius}
}
}
\usage{
raw_terra
}
\description{
CO2 and CH4 measured simultaneoulsy
}
\examples{
raw_terra
}
\keyword{datasets}
