% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-record_liahovden.R
\docType{data}
\name{record_liahovden}
\alias{record_liahovden}
\title{Measurements meta data at Liahovden}
\format{
A tibble with 138 rows and 3 variables
\describe{
\item{turfID}{Unique ID of the turf in which the measurement took place.}
\item{type}{Type of measurement: ecosystems respiration (ER)
or net ecosystem exchange (NEE).}
\item{start}{Datetime at which the measurement was started.}
}
}
\usage{
record_liahovden
}
\description{
Measurements meta data as recorded on the field at site Liahovden
}
\examples{
record_liahovden
}
\keyword{datasets}
