% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-record_short.R
\docType{data}
\name{record_short}
\alias{record_short}
\title{Measurements meta data}
\format{
A tibble with 6 rows and 3 variables
\describe{
\item{turfID}{Unique ID of the turf in which the measurement took place.}
\item{type}{Type of measurement: ecosystems respiration (ER)
or net ecosystem exchange (NEE).}
\item{start}{Datetime at which the measurement was started.}
}
}
\usage{
record_short
}
\description{
Measurements meta data as recorded on the field
}
\examples{
record_short
}
\keyword{datasets}
