% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-terra_record.R
\docType{data}
\name{terra_record}
\alias{terra_record}
\title{Field record}
\format{
A tibble with 12 rows and 1 variable
\describe{
\item{start}{Start datetime of each flux measurement}
}
}
\usage{
terra_record
}
\description{
Field record
}
\examples{
terra_record
}
\keyword{datasets}
