% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-twogases_record.R
\docType{data}
\name{twogases_record}
\alias{twogases_record}
\title{Two gases field record}
\format{
A tibble with 12 rows and 1 variable
\describe{
\item{start}{Start datetime of each flux measurement}
}
}
\usage{
twogases_record
}
\description{
Two gases field record
}
\examples{
twogases_record
}
\keyword{datasets}
