% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wet_conc.R
\docType{data}
\name{wet_conc}
\alias{wet_conc}
\title{CO2 and H2O concentration}
\format{
A tibble with 18 rows and 4 variables
\describe{
\item{Time}{Time in format hh:mm:ss}
\item{Date}{Date in format yyyy-mm-dd}
\item{co2}{CO2 concentration before wet air correction}
\item{h20}{H2O concentration before wet air correction}
}
}
\usage{
wet_conc
}
\description{
CO2 and H2O concentration measurements
}
\examples{
wet_conc
}
\keyword{datasets}
