% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_calc.R
\name{flux_calc}
\alias{flux_calc}
\title{to calculate a flux based on the rate of change of gas concentration over time}
\usage{
flux_calc(
  slope_df,
  slope_col,
  chamber_volume = 24.5,
  tube_volume = 0.075,
  atm_pressure = 1,
  plot_area = 0.0625,
  R_const = 0.082057,
  cols_keep = c(),
  cols_ave = c(),
  fluxID_col = "fluxID",
  temp_air_col = "temp_air",
  temp_air_unit = "celsius"
)
}
\arguments{
\item{slope_df}{dataframe of flux slopes}

\item{slope_col}{column containing the slope to calculate the flux}

\item{chamber_volume}{volume of the flux chamber in L, default for Three-D chamber (25x24.5x40cm)}

\item{tube_volume}{volume of the tubing in L, default for summer 2020 setup}

\item{atm_pressure}{atmoshperic pressure, assumed 1 atm}

\item{plot_area}{area of the plot in m^2, default for Three-D}

\item{R_const}{gas constant, in L\emph{atm}K^(-1)*mol^(-1)}

\item{cols_keep}{columns to keep from the input to the input. Those columns need to have unique values for each flux}

\item{cols_ave}{columns with values that should be averaged for each flux in the ouput. Not that NA are removed in mean calculation}

\item{fluxID_col}{column containing the fluxID}

\item{temp_air_col}{column containing the air temperature used to caculate fluxes. Will be averaged with NA removed.}

\item{temp_air_unit}{units for air temperature. Has to be either celsius, fahrenheit or kelvin}
}
\value{
a df containing fluxID, fluxes, temperature average for each flux, slope used for each flux calculation, and any columns specified in cols_keep and cols_ave
}
\description{
to calculate a flux based on the rate of change of gas concentration over time
}
\examples{
data(slopes0)
flux_calc(slopes0, slope_col = "slope_tz")
}
