% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_calc.R
\name{flux_calc}
\alias{flux_calc}
\title{calculates fluxes}
\usage{
flux_calc(
  slope_df,
  slope_col,
  cut_col = c(),
  keep_filter = c(),
  chamber_volume = 24.5,
  tube_volume = 0.075,
  atm_pressure = 1,
  plot_area = 0.0625,
  R_const = 0.082057,
  cols_keep = c(),
  cols_ave = c(),
  fluxID_col = "f_fluxID",
  temp_air_col = "temp_air",
  temp_air_unit = "celsius"
)
}
\arguments{
\item{slope_df}{dataframe of flux slopes}

\item{slope_col}{column containing the slope to calculate the flux}

\item{cut_col}{column containing cutting information}

\item{keep_filter}{name in cut_col of data to keep}

\item{chamber_volume}{volume of the flux chamber in L,
default for Three-D project chamber (25x24.5x40cm),
can also be a column in case it is a variable}

\item{tube_volume}{volume of the tubing in L, default for summer 2020 setup,
can also be a column in case it is a variable}

\item{atm_pressure}{atmoshperic pressure, assumed 1 atm,
can be a constant (numerical) or a variable (column name)}

\item{plot_area}{area of the plot in m^2, default for Three-D}

\item{R_const}{gas constant (0.082057 L\emph{atm}K^(-1)*mol^(-1))}

\item{cols_keep}{columns to keep from the input to the output.
Those columns need to have unique values for each flux.}

\item{cols_ave}{columns with values that should be averaged
for each flux in the ouput. Note that NA are removed in mean calculation}

\item{fluxID_col}{column containing the fluxID}

\item{temp_air_col}{column containing the air temperature used
to caculate fluxes. Will be averaged with NA removed.}

\item{temp_air_unit}{units in which air temperature was measured.
Has to be either celsius, fahrenheit or kelvin}
}
\value{
a df containing fluxID, fluxes (in mmol*m^(-2)*h^(-1)), temperature average for each flux,
slope used for each flux calculation,
and any columns specified in cols_keep and cols_ave.
}
\description{
calculates a flux based on the rate of change
of gas concentration over time
}
\examples{
data(slopes0)
flux_calc(slopes0, slope_col = "f_slope_tz")
}
