% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_conc.R
\name{flux_conc}
\alias{flux_conc}
\title{Transforms concentration from fractional to volumetric units}
\usage{
flux_conc(
  conc_df,
  f_conc,
  temp_air_col,
  atm_pressure = 1,
  f_fluxid = f_fluxid,
  temp_air_unit = "celsius"
)
}
\arguments{
\item{conc_df}{dataframe of gas concentration over time}

\item{f_conc}{column with gas concentration}

\item{temp_air_col}{column containing the air temperature used
to convert concentration.}

\item{atm_pressure}{atmospheric pressure in atm,
can be a constant (numerical) or a variable (column name).
Default is 1.}

\item{f_fluxid}{column with ID of each flux}

\item{temp_air_unit}{units in which air temperature was measured.
Has to be either \code{celsius} (default), \code{fahrenheit} or \code{kelvin.}}
}
\value{
The same dataframe, with additional column 'f_conc_vol' in volumetric
concentration.
}
\description{
This function converts gas concentrations from fractional units
(e.g., ppm) to volumetric units (e.g., μmol/liter) based on the ideal gas
law and the specific conditions of the measurement.
}
\details{
Required temperature and pressure data
This function requires that each gas concentration data points to
be paired with a corresponding air temperature and pressure measurement. If
missing, data will be filled in the downup order. Pressure can also be
provided as a constant for the entire dataset.

Units
The units of the newly calculated volumetric concentration follow the units
of the fractional concentration provided. For exemple, if the input is in
'ppm', the result is in 'µmol/L'; if the input is in 'ppb', the result is in
'nmol/L'; if the input is in 'ppt', the result is in 'pmol/L'.
}
\examples{
data(co2_conc)
flux_conc(co2_conc, conc, temp_air)
}
