% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fitting.R
\name{flux_fitting}
\alias{flux_fitting}
\title{Fitting a model to concentration data and estimating the slope}
\usage{
flux_fitting(
  conc_df,
  start_cut = 0,
  end_cut = 0,
  start_col = "f_start",
  end_col = "f_end",
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  fluxID_col = "f_fluxID",
  t_window = 20,
  Cz_window = 15,
  b_window = 10,
  a_window = 10,
  roll_width = 15,
  t_zero = 0,
  fit_type
)
}
\arguments{
\item{conc_df}{dataframe of gas concentration over time}

\item{start_cut}{time to discard at the start of the measurements
(in seconds)}

\item{end_cut}{time to discard at the end of the measurements (in seconds)}

\item{start_col}{column with datetime when the measurement started}

\item{end_col}{column with datetime when the measurement ended}

\item{datetime_col}{column with datetime of each concentration measurement}

\item{conc_col}{column with gas concentration data}

\item{fluxID_col}{column with ID of each flux}

\item{t_window}{enlarge focus window before and after tmin and tmax
(exponential fit)}

\item{Cz_window}{window used to calculate Cz, at the beginning of cut window
(exponential fit)}

\item{b_window}{window to estimate b. It is an interval after tz where
it is assumed that the model fits the data perfectly (exponential fit)}

\item{a_window}{window at the end of the flux to estimate a (exponential fit)}

\item{roll_width}{width of the rolling mean for CO2 when looking for tz,
ideally same as Cz_window (exponential fit)}

\item{t_zero}{time at which the slope should be calculated
(for quadratic fit)}

\item{fit_type}{exponential, quadratic or linear.
Exponential is using the exponential model from Zhao et al (2018)}
}
\value{
a dataframe with the slope at t zero,
and parameters of a model of gas concentration over time
}
\description{
fits gas concentration over time data with a model
(exponential, quadratic or linear) and provides the slope later used
to calculate gas fluxes with flux_calc
}
\examples{
data(co2_conc)
flux_fitting(co2_conc, fit_type = "exp")
flux_fitting(co2_conc, fit_type = "quadratic", t_zero = 10, end_cut = 30)
}
\references{
Zhao, P., Hammerle, A., Zeeman, M., Wohlfahrt, G., 2018.
On the calculation of daytime CO2 fluxes measured by automated closed
transparent chambers. Agricultural and Forest Meteorology 263, 267–275.
https://doi.org/10.1016/j.agrformet.2018.08.022
}
