% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fitting_lin.R
\name{flux_fitting_lin}
\alias{flux_fitting_lin}
\title{linear fit to gas concentration over time}
\usage{
flux_fitting_lin(
  conc_df,
  start_cut = 0,
  end_cut = 0,
  start_col = "start",
  end_col = "end",
  datetime_col = "datetime",
  conc_col = "conc",
  fluxID_col = "fluxID"
)
}
\arguments{
\item{conc_df}{dataframe of gas concentration over time}

\item{start_cut}{to cut at the start}

\item{end_cut}{to cut at the end, if you notice on the graphs that the match was not precise enough}

\item{start_col}{column with datetime when the measurement started}

\item{end_col}{column with datetime when the measurement ended}

\item{datetime_col}{column with datetime of each concentration measurement}

\item{conc_col}{column with gas concentration data}

\item{fluxID_col}{column with ID of each flux}
}
\value{
a df with the modelled gas concentration, slope, intercept, std error, r square and p value of the linear model
}
\description{
fits a linear model to the gas concentration over time
}
\examples{
data(co2_conc)
flux_fitting_lin(co2_conc)
}
