% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fitting_lm.R
\name{flux_fitting_lm}
\alias{flux_fitting_lm}
\title{linear fit to gas concentration over time}
\usage{
flux_fitting_lm(conc_df_cut, conc_df, f_conc, f_fluxid)
}
\arguments{
\item{conc_df_cut}{dataframe of gas concentration over time, cut}

\item{conc_df}{dataframe of gas concentration over time}

\item{f_conc}{column with gas concentration}

\item{f_fluxid}{column with ID of each flux}
}
\value{
a df with the modeled gas concentration, slope, intercept,
std error, r square and p value of the linear model
}
\description{
fits a linear model to the gas concentration over time
}
\keyword{internal}
