% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fitting_quadratic.R
\name{flux_fitting_quadratic}
\alias{flux_fitting_quadratic}
\title{quadratic fit to gas concentration over time}
\usage{
flux_fitting_quadratic(conc_df, start_cut, end_cut, t_zero)
}
\arguments{
\item{conc_df}{dataframe of gas concentration over time}

\item{start_cut}{time to discard at the start of the measurements
(in seconds)}

\item{end_cut}{time to discard at the end of the measurements (in seconds)}

\item{t_zero}{time at which the slope should be calculated}
}
\value{
a df with the modeled gas concentration, slope, intercept,
std error, r square and p value of the quadratic model
}
\description{
fits a quadratic model to the gas concentration over time
}
