% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fitting_segment.R
\name{flux_fitting_segment}
\alias{flux_fitting_segment}
\title{Fitting function while accounting for leaks}
\usage{
flux_fitting_segment(
  conc_df,
  start_cut,
  end_cut,
  par_col,
  h2o_col,
  signal_strength_col,
  h2o_correction,
  min_seg_length
)
}
\arguments{
\item{signal_strength_tresh}{Threshold for valid signal strength}

\item{par_thresh}{Photosynthetically Active Radiation (PAR) threshold}
}
\value{
a df with modeled gas concentration (NA outside of selected segments)
}
\description{
separate fluxes in segments of similar slopes with stable PAR
}
