% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_fun_check.R
\name{flux_fun_check}
\alias{flux_fun_check}
\title{checking that arguments and columns are in the correct class}
\usage{
flux_fun_check(df, col_numeric = c(), col_datetime = c(), arg_numeric = c())
}
\arguments{
\item{df}{a dataframe to check}

\item{col_numeric}{a vector of column names that have to be numeric}

\item{col_datetime}{a vector of column names that have to be time}

\item{arg_numeric}{a vectore of arguments that have to be numeric}
}
\description{
checking that arguments and columns are in the correct class
}
