% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_gep.R
\name{flux_gep}
\alias{flux_gep}
\title{Calculates GEP}
\usage{
flux_gep(
  fluxes_df,
  id_cols,
  flux_col,
  type_col,
  datetime_col,
  par_col,
  nee_arg = "NEE",
  er_arg = "ER",
  cols_keep = c()
)
}
\arguments{
\item{fluxes_df}{a dataframe containing NEE and ER}

\item{id_cols}{columns used to identify each pair of ER and NEE}

\item{flux_col}{column containing flux values}

\item{type_col}{column containing type of flux (NEE or ER)}

\item{datetime_col}{column containing start of measurement as datetime}

\item{par_col}{column containing PAR values for each flux}

\item{nee_arg}{argument designating NEE fluxes in type column}

\item{er_arg}{argument designating ER fluxes in type column}

\item{cols_keep}{columns to keep from fluxes_df. Values from NEE row will be
filled in GEP row.}
}
\value{
a df with GEP as a flux type, with PAR and datetime from the NEE
measurement for each pair of ER and NEE
}
\description{
to calculate gross ecosystem production (GEP) from net ecosystem
(NEE) exchange and ecosystem respiration (ER) as GEP = NEE - ER
}
\examples{
data(co2_fluxes)
flux_gep(co2_fluxes, id_cols = "turfID", flux_col = "flux",
type_col = "type", datetime_col = "f_start", par_col = "PAR",
cols_keep = c("temp_soil"))
}
