% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_param_exp.R
\name{flux_param_exp}
\alias{flux_param_exp}
\title{prepares text to print for flux_plot function}
\usage{
flux_param_exp(slopes_df, f_datetime)
}
\arguments{
\item{slopes_df}{the slopes_df that is being provided to flux_plot}

\item{f_datetime}{column with datetime of each data point}
}
\description{
creates a df with quality flags and quality diagnostics
to print on the plots produced by flux_plot.
flux_param_lm is for fit in the lm family (linear and quadratic)
flux_param_exp is for the exponential fit
}
\keyword{internal}
