% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot.R
\name{flux_plot}
\alias{flux_plot}
\title{ploting fluxes for fit evaluation}
\usage{
flux_plot(
  slopes_df,
  fit_type,
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  cut_col = "f_cut",
  fit_col = "f_fit",
  fit_slope_col = "f_fit_slope",
  quality_flag_col = "f_quality_flag",
  fluxID_col = "f_fluxID",
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  start_col = "f_start",
  b_col = "f_b",
  cor_coef_col = "f_cor_coef",
  RMSE_col = "f_RMSE",
  f_date_breaks = "1 min",
  f_minor_breaks = "10 sec",
  f_date_labels = "\%e/\%m \\n \%H:\%M",
  f_ylim_upper = 800,
  f_ylim_lower = 400,
  f_scales = "free",
  f_plotname = "plot_quality",
  f_ncol = 4,
  f_nrow = 3,
  print_plot = "FALSE"
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{fit_type}{model used in flux_fitting, exponential or linear}

\item{datetime_col}{column containing datetime of
each concentration measurement}

\item{conc_col}{column containing gas concentration data}

\item{cut_col}{column containing cut factor from the
flux_fitting function ("cut" or "keep")}

\item{fit_col}{column containing the modelled fit of the flux}

\item{fit_slope_col}{column containing the modelled slope at tz}

\item{quality_flag_col}{column containing the flags produced by flux_quality}

\item{fluxID_col}{column containing unique IDs for each flux}

\item{pvalue_col}{column containing the p-value of each flux}

\item{rsquared_col}{column containing the r squared to be used
for the quality assessment}

\item{start_col}{column containing the datetime of the start of each flux}

\item{b_col}{column containing the b parameter of the exponential fit}

\item{cor_coef_col}{column containing the correlation coefficient
produced by flux_quality}

\item{RMSE_col}{column containing the RMSE produced by flux_quality}

\item{f_date_breaks}{date_breaks argument for scale_x_datetime}

\item{f_minor_breaks}{minor breaks argument for scale_x_datetime}

\item{f_date_labels}{date_labels argument for scale_x_datetime}

\item{f_ylim_upper}{y axis upper limit}

\item{f_ylim_lower}{y axis lower limit}

\item{f_scales}{argument for scales in facet_wrap ("fixed" or "free")}

\item{f_plotname}{filename for the extracted pdf file}

\item{f_ncol}{ncol argument for facet_wrap}

\item{f_nrow}{number of row per page in extracted pdf file}

\item{print_plot}{FALSE or TRUE, if TRUE it prints the plot in R
but will take time depending on the size of the dataset}
}
\description{
plots the fluxes and indicates what should be discarded
or replaced by zero
}
\examples{
data(slopes0_flag)
flux_plot(slopes0_flag, fit_type = "exp",
fit_slope_col = "f_fit_slope", print_plot = TRUE)
}
