% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_exp.R
\name{flux_plot_exp}
\alias{flux_plot_exp}
\title{plotting fluxes with exponential fit}
\usage{
flux_plot_exp(slopes_df, conc_col, datetime_col, y_text_position)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{conc_col}{column with gas concentration}

\item{datetime_col}{column with datetime of each data point}

\item{y_text_position}{position of the text box}
}
\description{
plots the fluxes that were fitted with
an exponential model
}
