% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_exp.R
\name{flux_plot_exp}
\alias{flux_plot_exp}
\title{plotting fluxes with exponential fit}
\usage{
flux_plot_exp(
  slopes_df,
  cut_arg = "cut",
  fit_slope_col = "f_fit_slope",
  b_col = "f_b",
  cor_coef_col = "f_cor_coef",
  rmse_col = "f_RMSE",
  y_text_position = 500
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{cut_arg}{argument pointing rows to be cut from the measurements}

\item{fit_slope_col}{column containing the modeled slope at tz}

\item{b_col}{column containing the b parameter of the exponential fit}

\item{cor_coef_col}{column containing the correlation coefficient
produced by flux_quality}

\item{rmse_col}{column containing the RMSE produced by flux_quality}

\item{y_text_position}{position of the text box}
}
\description{
plots the fluxes that were fitted with
an exponential model
}
