% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_exp.R
\name{flux_plot_exp}
\alias{flux_plot_exp}
\title{plotting fluxes with exponential fit}
\usage{
flux_plot_exp(slopes_df, f_conc, f_datetime, y_text_position, kappamax)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{f_conc}{column with gas concentration}

\item{f_datetime}{column with datetime of each data point}

\item{y_text_position}{position of the text box}
}
\description{
plots the fluxes that were fitted with
an exponential model
}
