% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_exp.R
\name{flux_plot_exp}
\alias{flux_plot_exp}
\title{plotting fluxes with exponential fit}
\usage{
flux_plot_exp(slopes_df, cut_arg, y_text_position = 500)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{cut_arg}{argument pointing rows to be cut from the measurements}

\item{y_text_position}{position of the text box}
}
\description{
plots the fluxes that were fitted with
an exponential model
}
