% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_flag.R
\name{flux_plot_flag}
\alias{flux_plot_flag}
\title{creates the flag column to be used by flux_plot}
\usage{
flux_plot_flag(slopes_df, param_df)
}
\arguments{
\item{slopes_df}{as provided in flux_plot}

\item{param_df}{as provided by flux_param}
}
\description{
creates a column with quality flags (from flux_quality)
for the part of the rows to be kept, and cut flag for rows to be discarded
}
