% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_lin.R
\name{flux_plot_lin}
\alias{flux_plot_lin}
\title{plotting fluxes with linear fit}
\usage{
flux_plot_lin(slopes_df, y_text_position = 500, cut_arg = "cut")
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{y_text_position}{position of the text box}

\item{cut_arg}{argument pointing rows to be cut from the measurements}
}
\description{
plots the fluxes that were fitted
with a linear model
}
