% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_lin.R
\name{flux_plot_lin}
\alias{flux_plot_lin}
\title{ploting fluxes for fit evaluation}
\usage{
flux_plot_lin(
  slopes_df,
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  cut_col = "f_cut",
  fit_col = "f_fit",
  quality_flag_col = "f_quality_flag",
  fluxID_col = "f_fluxID",
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  start_col = "f_start",
  color_discard = "#D55E00",
  color_cut = "#D55E00",
  color_keep = "#009E73",
  color_ok = "#000000",
  color_zero = "#CC79A7",
  f_date_breaks = "1 min",
  f_minor_breaks = "10 sec",
  f_date_labels = "\%e/\%m \\n \%H:\%M",
  f_ylim_upper = 800,
  f_ylim_lower = 400,
  f_scales = "free",
  f_plotname = "plot_quality_lin",
  f_ncol = 4,
  f_nrow = 3,
  y_text_position = 500,
  f_nudge_y = 100,
  print_plot = "FALSE"
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{datetime_col}{column containing datetime of
each concentration measurement}

\item{conc_col}{column containing gas concentration data}

\item{cut_col}{column containing cut factor from the
flux_fitting function ("cut" or "keep")}

\item{fit_col}{column containing the modelled fit of the flux}

\item{quality_flag_col}{column containing the flags produced by flux_quality}

\item{fluxID_col}{column containing unique IDs for each flux}

\item{pvalue_col}{column containing the p-value of each flux}

\item{rsquared_col}{column containing the r squared
used for the quality assessment}

\item{start_col}{column containing the datetime of the start of each flux}

\item{color_discard}{color for fits with a discard quality flag}

\item{color_cut}{color for the part of the flux that is cut}

\item{color_keep}{color for the part of the flux that is kept}

\item{color_ok}{color for fits with an ok quality flag}

\item{color_zero}{color for fits with a zero quality flag}

\item{f_date_breaks}{date_breaks argument for scale_x_datetime}

\item{f_minor_breaks}{minor breaks argument for scale_x_datetime}

\item{f_date_labels}{date_labels argument for scale_x_datetime}

\item{f_ylim_upper}{y axis upper limit}

\item{f_ylim_lower}{y axis lower limit}

\item{f_scales}{argument for scales in facet_wrap ("fixed" or "free")}

\item{f_plotname}{filename for the extracted pdf file}

\item{f_ncol}{ncol argument for facet_wrap}

\item{f_nrow}{number of row per page in extracted pdf file}

\item{y_text_position}{position of the text box}

\item{f_nudge_y}{to nudge the text box with the parameters above the modelled flux}

\item{print_plot}{FALSE or TRUE, if TRUE it prints the plot in R
but will take time depending on the size of the dataset}
}
\description{
plots the fluxes and indicates what should be
discarded or replaced by zero
}
