% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_lin.R
\name{flux_plot_lin}
\alias{flux_plot_lin}
\title{plotting fluxes with linear fit}
\usage{
flux_plot_lin(slopes_df, f_conc, f_datetime, y_text_position)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{f_conc}{column with gas concentration}

\item{f_datetime}{column with datetime of each data point}

\item{y_text_position}{position of the text box}
}
\description{
plots the fluxes that were fitted
with a linear model
}
\keyword{internal}
