% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_lin.R
\name{flux_plot_lin}
\alias{flux_plot_lin}
\title{ploting fluxes for fit evaluation}
\usage{
flux_plot_lin(
  slopes_df,
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  y_text_position = 500,
  cut_arg = "cut"
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{pvalue_col}{column containing the p-value of each flux}

\item{rsquared_col}{column containing the r squared
used for the quality assessment}

\item{y_text_position}{position of the text box}

\item{cut_arg}{argument pointing rows to be cut from the measurements
but will take time depending on the size of the dataset}
}
\description{
plots the fluxes and indicates what should be
discarded or replaced by zero
}
