% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_plotly.R
\name{flux_plot_plotly}
\alias{flux_plot_plotly}
\title{print plots in pdf}
\usage{
flux_plot_plotly(f_plot, f_plotname, nb_fluxid, plotly_args)
}
\arguments{
\item{f_plot}{plot from flux_plot}

\item{f_plotname}{name of plot}

\item{nb_fluxid}{number of facets}

\item{plotly_args}{arguments for plotly in the form
\code{list(ncol, width, ratio)}}
}
\description{
print plots in pdf
}
