% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_quadratic.R
\name{flux_plot_quadratic}
\alias{flux_plot_quadratic}
\title{ploting fluxes for fit evaluation}
\usage{
flux_plot_quadratic(
  slopes_df,
  datetime_col = "f_datetime",
  conc_col = "f_conc",
  cut_col = "f_cut",
  fit_col = "f_fit",
  quality_flag_col = "f_quality_flag",
  fluxID_col = "f_fluxID",
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  start_col = "f_start",
  y_text_position = 500
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{datetime_col}{column containing datetime of
each concentration measurement}

\item{conc_col}{column containing gas concentration data}

\item{cut_col}{column containing cut factor from the
flux_fitting function ("cut" or "keep")}

\item{fit_col}{column containing the modelled fit of the flux}

\item{quality_flag_col}{column containing the flags produced by flux_quality}

\item{fluxID_col}{column containing unique IDs for each flux}

\item{pvalue_col}{column containing the p-value of each flux}

\item{rsquared_col}{column containing the r squared
used for the quality assessment}

\item{start_col}{column containing the datetime of the start of each flux}

\item{y_text_position}{position of the text box
but will take time depending on the size of the dataset}
}
\description{
plots the fluxes and indicates what should be
discarded or replaced by zero
}
