% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_quadratic.R
\name{flux_plot_quadratic}
\alias{flux_plot_quadratic}
\title{plotting fluxes with a quadratic fit}
\usage{
flux_plot_quadratic(
  slopes_df,
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  y_text_position = 500,
  cut_arg = "cut"
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{pvalue_col}{column containing the p-value of each flux}

\item{rsquared_col}{column containing the r squared of each flux}

\item{y_text_position}{position of the text box}

\item{cut_arg}{argument pointing rows to be cut from the measurements}
}
\description{
specific part of flux_plot for
quadratic fit
}
