% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_plot_segment.R
\name{flux_plot_segment}
\alias{flux_plot_segment}
\title{plots the fluxes as part of flux_plot, with flags from flux_quality and modelled value from flux_segment}
\usage{
flux_plot_segment(slopes_df, y_text_position = 500, cut_arg = "cut")
}
\description{
plots the fluxes as part of flux_plot, with flags from flux_quality and modelled value from flux_segment
}
