% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_exp.R
\name{flux_quality_exp}
\alias{flux_quality_exp}
\title{assessing fluxes quality calculated with an exponential fit
quality assessment for the fluxes calculated with the exponential model}
\usage{
flux_quality_exp(
  slopes_df,
  slope_col = "f_slope_tz",
  weird_fluxesID = c(),
  b_col = "f_b",
  RMSE_threshold = 25,
  cor_threshold = 0.5,
  b_threshold = 1
)
}
\arguments{
\item{slopes_df}{dataset containing slopes, fluxID,
and parameters of the exponential expression}

\item{slope_col}{column containing the slope of the exponential expression
used for the calculation}

\item{weird_fluxesID}{vector of fluxIDs that should be discarded
by the user's decision}

\item{b_col}{column containing the b parameter of the exponential expression}

\item{RMSE_threshold}{threshold for the RMSE of each flux above
which the fit is considered unsatisfactory}

\item{cor_threshold}{threshold for the correlation coefficient
of gas concentration with time below which the correlation
is considered non significant}

\item{b_threshold}{threshold for the b parameter. Defines a window
with its opposite inside which the fit is considered good enough.}

\item{fluxID_col}{column containing unique IDs for each flux}

\item{conc_col}{column containing the measured gas concentration}

\item{time_col}{column containing the time of each measurement in seconds}

\item{fit_col}{column containing the modeled data}

\item{cut_col}{column containing the cutting information}

\item{ambient_conc}{ambient gas concentration in ppm at the site of
measurement (used to detect measurement that started with a polluted setup)}

\item{error}{error of the setup, defines a window in which
the starting values are considered acceptable}
}
\value{
same dataframe with added flag and corrected slopes columns
}
\description{
indicates if fluxes should be discarded or replaced
by 0 according to parameters set by user
}
