% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_exp.R
\name{flux_quality_exp}
\alias{flux_quality_exp}
\title{assessing fluxes quality calculated with an exponential fit
quality assessment for the fluxes calculated with the exponential model}
\usage{
flux_quality_exp(
  slopes_df,
  ambient_conc = 421,
  error = 100,
  fluxID_col = "f_fluxID",
  slope_col = "f_slope",
  weird_fluxesID = c(),
  conc_col = "f_conc",
  b_col = "f_b",
  time_col = "f_time",
  fit_col = "f_fit",
  cut_col = "f_cut",
  RMSE_threshold = 25,
  cor_threshold = 0.5,
  b_threshold = 1
)
}
\arguments{
\item{slopes_df}{dataset containing slopes, fluxID, and parameters of the exponential expression}

\item{ambient_conc}{ambient gas concentration in ppm at the site of measurement (used to detect measurement that started with a polluted setup)}

\item{error}{error of the setup, defines a window in which the starting values are considered acceptable}

\item{fluxID_col}{column containing unique IDs for each flux}

\item{slope_col}{column containing the slope of the exponential expression used for the calculation}

\item{weird_fluxesID}{vector of fluxIDs that should be discarded by the user's decision}

\item{conc_col}{column containing the measured gas concentration}

\item{b_col}{column containing the b parameter of the exponential expression}

\item{time_col}{column containing the time of each measurement in seconds}

\item{fit_col}{column containing the modeled data}

\item{cut_col}{column containing the cutting information}

\item{RMSE_threshold}{threshold for the RMSE of each flux above which the fit is considered unsatisfactory}

\item{cor_threshold}{threshold for the correlation coefficient of gas concentration with time below which the correlation is considered non significant}

\item{b_threshold}{threshold for the b parameter. Defines a window with its opposite inside which the fit is considered good enough.}
}
\value{
same dataframe with added flag and corrected slopes columns
}
\description{
indicates if fluxes should be discarded or replaced by 0 according to parameters set by user
}
\examples{
data(slopes0)
flux_quality_exp(slopes0,
fluxID_col = "fluxID",
conc_col = "conc",
b_col = "b",
time_col = "time",
fit_col = "fit",
slope_col = "slope_tz",
cut_col = "cut"
)
}
