% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_lin.R
\name{flux_quality_lin}
\alias{flux_quality_lin}
\title{to provide quality assessment for the fluxes calculated with the linear model}
\usage{
flux_quality_lin(
  fluxes_df,
  pvalue_threshold = 0.3,
  rsquared_threshold = 0.7,
  blind_replace = FALSE,
  fluxID_col = "fluxID",
  fluxes_col = "fluxes",
  pvalue_col = "p.value",
  rsquared_col = "r.squared"
)
}
\description{
to provide quality assessment for the fluxes calculated with the linear model
}
