% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_lin.R
\name{flux_quality_lin}
\alias{flux_quality_lin}
\title{quality assessment for the fluxes calculated with the linear model}
\usage{
flux_quality_lin(
  slopes_df,
  ambient_conc = 421,
  error = 100,
  fluxID_col = "f_fluxID",
  slope_col = "f_slope",
  conc_col = "f_conc",
  weird_fluxesID = c(),
  pvalue_col = "f_pvalue",
  rsquared_col = "f_rsquared",
  pvalue_threshold = 0.3,
  rsquared_threshold = 0.7
)
}
\arguments{
\item{slopes_df}{dataset containing slopes, fluxID, p.value and r.squared}

\item{ambient_conc}{ambient gas concentration in ppm at the site of
measurement (used to detect measurement that started with a polluted setup)}

\item{error}{error of the setup, defines a window in
which the starting values are considered acceptable}

\item{fluxID_col}{column containing unique IDs for each flux}

\item{slope_col}{column containing the slope of each flux
(as calculated by the flux_fitting function)}

\item{conc_col}{column containing the measured gas concentration}

\item{weird_fluxesID}{vector of fluxIDs that should be discarded
by the user's decision}

\item{pvalue_col}{column containing the p-value of each flux}

\item{rsquared_col}{column containing the r squared to be used for
the quality assessment}

\item{pvalue_threshold}{threshold of p-value below which the change
of gas concentration over time is considered not significant (user decided)}

\item{rsquared_threshold}{threshold of r squared value below which
the linear model is considered an unsatisfactory fit}
}
\value{
same dataframe with added flag and corrected slopes columns
}
\description{
indicates if fluxes should be discarded or replaced by 0
according to parameters set by user
}
\examples{
data(slopes0lin)
flux_quality_lin(slopes0lin)
}
