% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_quality_segment.R
\name{flux_quality_segment}
\alias{flux_quality_segment}
\title{Quality control for segmented measurements}
\usage{
flux_quality_segment(
  slopes_df,
  pvalue_col,
  rsquared_col,
  f_flag_fit_col,
  par_col,
  sign_str_col,
  par_threshold,
  sign_str_threshold,
  pvalue_threshold,
  rsquared_threshold,
  sd_threshold,
  cut_arg,
  force_discard,
  force_ok
)
}
\arguments{
\item{slopes_df}{dataset containing slopes}

\item{pvalue_col}{column containing the p-value of each flux
(linear, quadratic, segment)}

\item{rsquared_col}{column containing the r squared of each flux
(linear, quadratic, segment)}

\item{f_flag_fit_col}{column flagging measurements that were too short to
find segments (optional; provided by \link[flux_fitting]{flux_fitting}).}

\item{par_col}{column containing segment average PAR data}

\item{sign_str_col}{column containing segment average signal strength}

\item{par_threshold}{PAR value threshold under which a segment should be
discarded (if PAR data are not provided, it will just be ignored)}

\item{sign_str_threshold}{signal strength threshold under which a segment
should be discarded
(if signal strength is not provided it will just be ignored)}

\item{pvalue_threshold}{threshold of p-value below which the change of
gas concentration over time is considered not significant
(linear, quadratic, segment)}

\item{rsquared_threshold}{threshold of r squared value below which
the linear model is considered an unsatisfactory fit
(linear, quadratic, segment)}

\item{sd_threshold}{standard deviation threshold under which a measurement
should be discarded. The standard deviation is calculated as
\link[=sqrt(sum(f_segment_length * (f_slope_corr - f_mean_slope)^2) /
(((nb_segments_ok - 1) * sum(f_segment_length) / nb_segments_ok))]{sqrt(sum(f_segment_length * (f_slope_corr - f_mean_slope)^2) / (((nb_segments_ok - 1) * sum(f_segment_length) / nb_segments_ok))}
where 'f_segment_length' is the length of each segment;
'f_slope_corr' the slope of each segment after quality assessment
based on R² and p-value;
'f_mean_slope' the mean of the slope for the entire measurement
weighed with the length of each segment;
'nb_segments_ok' the number of segments, excluding those discarded based on
R² and p-value, in each segment.
The full equation is described in Smooth (1997).}

\item{cut_arg}{argument defining that the data point should be cut out}

\item{force_discard}{vector of fluxIDs that should be discarded
by the user's decision}

\item{force_ok}{vector of fluxIDs for which the user wants to keep
the calculated slope despite a bad quality flag}
}
\description{
provides quality flags for the segments and entire measurements
following the same standards as for the linear model
}
\references{
SMOOTH, Y. DATAPLOT Reference Manual, 1997 2-66
}
