% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-flux_tent_output.R
\docType{data}
\name{flux_tent_output}
\alias{flux_tent_output}
\title{CO2 concentration}
\format{
\describe{
\item{}{}
}
}
\usage{
flux_tent_output
}
\description{
CO2 concentration with measurements meta data, measured with a tent setup
}
\examples{
flux_tent_output
}
\keyword{datasets}
