% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_units.R
\name{flux_units}
\alias{flux_units}
\title{Returns a unit conversion coefficient for flux_calc}
\usage{
flux_units(
  flux_units,
  amount_units = c("mol", "mmol", "umol", "nmol", "pmol"),
  time_units = c("d", "h", "mn", "s"),
  surface_units = c("m2", "dm2", "cm2")
)
}
\arguments{
\item{flux_units}{desired units for the calculated fluxes}
}
\description{
Returns a unit conversion coefficient for flux_calc
}
