% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pftc7_segmented_short.R
\docType{data}
\name{pftc7_segmented_short}
\alias{pftc7_segmented_short}
\title{CO2 concentration}
\format{
a tibble with 26 columns and 1574 rows
\describe{
\item{co2_conc}{}
\item{h2o_conc}{}
\item{temperature_c}{}
\item{pressure_kpa}{}
\item{signal_strength}{}
\item{date_time}{}
\item{start_time}{}
\item{file_name}{}
\item{site}{}
\item{elevation}{}
\item{aspect}{}
\item{plot}{}
\item{day_night}{}
\item{measurement}{}
\item{redo}{}
\item{plot_id}{}
\item{par}{}
\item{f_fit}{}
\item{f_slope}{}
\item{f_time}{}
\item{f_rsq}{}
\item{f_rsq_adj}{}
\item{f_pval}{}
\item{f_segment_id}{}
\item{corrected_for_water_vapor}{}
\item{f_cut}{}
}
}
\usage{
pftc7_segmented_short
}
\description{
CO2 concentration with measurements meta data, fitted with segmentation tool
}
\examples{
pftc7_segmented_short
}
\keyword{datasets}
